Ensure clarity of index schemes within the nav function
There are a number of different indexing schemes being used within the
navigation file. With both 1 and 0 indexing of rows in neovim functions,
while there is also the 1 indexing of lua. This clarifies where each of
the different indexing patterns are being used and when we are
converting between each, fixing some off-by-one errors along the way.